+++ src/common/enums/store.ts
@@ -1,0 +1,3 @@
+export const STORE_KEY_PUSH = '__PUSH'
+
+export const STORE_KEY_VIEW_MODE = '__VIEW_MODE'