+++ src/components/Comment/DropdownActions/CollapseComment/index.tsx
@@ -1,7 +1,9 @@
 import Button from './Button'
 import Dialog from './Dialog'
 
-export default {
+const CollapseComment = {
   Dialog,
   Button
 }
+
+export default CollapseComment