+++ src/components/Layout/SideNav/index.tsx
@@ -33,8 +33,9 @@ const SideNav = () => {
   const isInNotification = router.pathname === PATHS.ME_NOTIFICATIONS.href
   const isInSearch = router.pathname === PATHS.SEARCH.href
   const isInMe =
-    router.asPath.indexOf('/me') >= 0 ||
-    router.query.userName === viewerUserName
+    router.asPath !== PATHS.ME_NOTIFICATIONS.as &&
+    (router.asPath.indexOf('/me') >= 0 ||
+      router.query.userName === viewerUserName)
   const isInDraftDetail = router.pathname === PATHS.ME_DRAFT_DETAIL.href
 
   return (
@@ -100,6 +101,7 @@ const SideNav = () => {
             placement="right-start"
             distance={16}
             offset={-16}
+            boundary="viewport"
             appendTo={process.browser ? document.body : undefined}
             zIndex={Z_INDEX.OVER_GLOBAL_HEADER}
             onShown={i => {
@@ -112,6 +114,7 @@ const SideNav = () => {
               activeIcon={<MeAvatar user={viewer} active />}
               active={isInMe}
               isMediumUp={isMediumUp}
+              canScrollTop={false}
               aira-haspopup="true"
             />
           </Dropdown>