+++ src/components/Layout/index.tsx
@@ -27,6 +27,7 @@ export const Layout: React.FC<LayoutProps> & {
   const isSmallUp = useResponsive('sm-up')
   const isLargeUp = useResponsive('lg-up')
   const router = useRouter()
+
   const mainClass = classNames({
     'l-col-4 l-col-sm-8 l-col-md-9 l-col-lg-9': true,
     [`bg-${bgColor}`]: !!bgColor,
@@ -43,10 +44,14 @@ export const Layout: React.FC<LayoutProps> & {
     'l-col-4 l-col-sm-7 l-offset-sm-1 l-col-md-7 l-offset-md-2 l-col-lg-3 l-offset-lg-0': true,
     'u-lg-down-hide': !asideShowInMobile
   })
+
   const isInSearch = router.pathname === PATHS.SEARCH.href
+  const isInDraftDetail = router.pathname === PATHS.ME_DRAFT_DETAIL.href
+
   const showNav = isSmallUp
   const showAside = isLargeUp || asideShowInMobile
   const showSearchBar = isLargeUp && !isInSearch
+  const showNavBar = !isSmallUp && !isInDraftDetail
 
   return (
     <>
@@ -80,9 +85,11 @@ export const Layout: React.FC<LayoutProps> & {
         )}
       </div>
 
-      <footer className="u-sm-up-hide">
-        <NavBar />
-      </footer>
+      {showNavBar && (
+        <footer className="u-sm-up-hide">
+          <NavBar />
+        </footer>
+      )}
 
       <style jsx>{styles}</style>
     </>