+++ src/components/Notice/SubscribedArticleNewCommentNotice.tsx
@@ -25,7 +25,7 @@ const SubscribedArticleNewCommentNotice = ({
   }
 
   const actorsCount = notice.actors.length
-  const isMultiActors = notice.actors && actorsCount > 1
+  const isMultiActors = actorsCount > 1
 
   return (
     <section className="container">
@@ -42,7 +42,7 @@ const SubscribedArticleNewCommentNotice = ({
           {notice.actors.slice(0, 2).map((actor, index) => (
             <Fragment key={index}>
               <NoticeActorName user={actor} />
-              {index < actorsCount - 1 && <span>、</span>}
+              {isMultiActors && index < 1 && <span>、</span>}
             </Fragment>
           ))}{' '}
           {isMultiActors && (