+++ src/views/ArticleDetail/AppreciationButton/CivicLikerButton.tsx
@@ -4,22 +4,15 @@ import CivicLikerDialog from './CivicLikerDialog'
 const CivicLikerButton = ({
   onClose,
   count,
-  total,
-  inFixedToolbar
+  total
 }: {
   onClose: () => void
   count?: number
   total: number
-  inFixedToolbar?: boolean
 }) => (
   <CivicLikerDialog onClose={onClose}>
     {({ open }) => (
-      <AppreciateButton
-        onClick={open}
-        count={count}
-        total={total}
-        inFixedToolbar={inFixedToolbar}
-      />
+      <AppreciateButton onClick={open} count={count} total={total} />
     )}
   </CivicLikerDialog>
 )